Remove stale finalize helper adapters
