+++ src/adapters/n8n/contracts/final-runtime-payload.ts
@@ -1,6 +1,36 @@
 import type { DebugContext } from '../../../lib/debug-context.js';
 import type { BriefJson } from '../../../types/brief.js';
-import type { EditorialBoundaryPayload } from './editorial-input.js';
+import type { BriefContext } from '../../../domain/editorial/resolution/resolve-editorial.js';
+
+export type GroqChoice = {
+  message?: {
+    content?: string;
+  };
+};
+
+export type GeminiDiagnosticsFields = {
+  geminiResponse?: unknown;
+  geminiRawPayload?: unknown;
+  geminiInspire?: unknown;
+  geminiStatusCode?: unknown;
+  geminiFinishReason?: unknown;
+  geminiCandidateCount?: unknown;
+  geminiResponseByteLength?: unknown;
+  geminiResponseTruncated?: unknown;
+  geminiExtractionPath?: unknown;
+  geminiTopLevelKeys?: unknown;
+  geminiPayloadKeys?: unknown;
+  geminiPartKinds?: unknown;
+  geminiExtractedTextLength?: unknown;
+  geminiPromptTokenCount?: unknown;
+  geminiCandidatesTokenCount?: unknown;
+  geminiTotalTokenCount?: unknown;
+  geminiThoughtsTokenCount?: unknown;
+};
+
+export type EditorialBoundaryPayload = BriefContext & GeminiDiagnosticsFields & {
+  choices?: GroqChoice[];
+};
 
 export type LongRangeCandidateLike = {
   name?: string;
